% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cor_test.R
\name{Cor_test}
\alias{Cor_test}
\title{Computed the RV coefficient between the metrics.}
\usage{
Cor_test(Varia_paysage_multi, method_cor)
}
\arguments{
\item{Varia_paysage_multi}{Tab generated by formodel()}

\item{method_cor}{correlation method, spearman or pearson}

\item{dist}{Vector of scales you choosed during you analusis in Chloe}
}
\value{
Return matrix of correlation between all metrics.
}
\description{
`Cor_test()` Return matrix of correlation between all metrics.
}
\details{
This function is meant to be used with the tab generated by formultivariate()
The RV coefficient is meant to compute the correlation between two tabs.
In this case metric's values at each scale are considered as tab. So a RV coefficient is computed
for each pairwise metrics with an associeted p-value.
}
