% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_multi.R
\name{launch_multi}
\alias{launch_multi}
\title{Function is to launch the GUI of the package, based on Shiny.}
\usage{
launch_multi()
}
\description{
`launch_multi()` If used on Rstudio the function will launch an interface to run
the analysis graphically.
}
\author{
Pierre-Gilles Lemasle <pg.lemasle@gmail.com>
}
