% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_find.R
\name{max_find}
\alias{max_find}
\title{Search peak in a vector}
\usage{
max_find(test)
}
\arguments{
\item{test}{vector including peak to find}

\item{temp}{The tab generated by formodel() function.}
}
\value{
return bolean vector where true indicate a peak
}
\description{
`max_find()` search the peak in a vector
}
\author{
Pierre-Gilles Lemasle <pg.lemasle@gmail.com>
}
