% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs.R
\name{plot_obs}
\alias{plot_obs}
\title{Graphically analysis of the relionship of abundance and metric value computed by Chloe at one scale
or the relionship of metric value and scales.}
\usage{
plot_obs(temp, dist = NULL, metrics, carab = NULL, sp_name = NULL,
  dist_col = NULL)
}
\arguments{
\item{temp}{The tab generated by formodel() function.}

\item{dist}{Vector of scales you choosed in Chloe when you created the ascii whith Cloe}

\item{metrics}{Vector of metrics of 1 or more you choosed in Chloe when you created the ascii whith Cloe}

\item{carab}{Optionnal argument containing your tab of species sampling, the tab must included a "id"
corresponding of the name of the sampling points?}

\item{sp_name}{Optionnal argument containing the name of the species you want to see analyse, the name
must match its column name contained in the "carab" tab}

\item{dist_col}{Optionnal argument containing the scale you want to analyse.}
}
\value{
Return a list of plot(s) of the length of "metrics".
}
\description{
`plot_obs()` Return a plot of of the relionship abundance of "sp_name" contained in "carab" and the metric(s)
in "metrics" if arguments "carab", "sp_name" and "dist_col" are given. Overwise return a plot of the relationship
between the value of the metric and the scales in dist.
}
\author{
Pierre-Gilles Lemasle <pg.lemasle@gmail.com>
}
