% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Chloe.properties.R
\name{read.Chloe.properties}
\alias{read.Chloe.properties}
\title{Reading the proprieties file generated by Chloe}
\usage{
read.Chloe.properties(path, type)
}
\arguments{
\item{path}{The path to your proprieties file generated in Chloe for your analysis}

\item{type}{Argument specifying if you want to read the "metrics" or the "Distance" used for
your analysis}
}
\value{
Return a vector of the metrics or scales used in Chloe in your analysis.
}
\description{
`read.Chloe.properties()` Return a vector of the metrics or the scales you used in Chloe for your
analysis
}
\details{
This function is made to be used to generate a tab used after in results_lm(), results_gam(),results_MARS,
 results_ALL() and tab_for_model().
}
\author{
Pierre-Gilles Lemasle <pg.lemasle@gmail.com>
}
