% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_GLM.R
\name{result_GLM}
\alias{result_GLM}
\title{Computed the R adjusted of the abundance/metric value models at each scales with GLM}
\usage{
result_GLM(x, dist, metrics, Sp_Name, tab = F,
  family_choice = poisson())
}
\arguments{
\item{dist}{Vector of scales you choosed during you analusis in Chloe}

\item{metrics}{Vector of one metric or more you choosed in Chloe during your analysis in Chloe}

\item{tab}{Argument "T" or "F" defining if you want to return plot (F) or a tab (T).}

\item{temp}{The tab generated by formodel() function merged by the id of sampling site
with the tab of your abundance sampling.}

\item{sp_name}{Optionnal argument containing the name of the species you want to see analyse, the name
must match its column name contained in the "carab" tab}
}
\value{
Return a list of plot(s) of the length of "metrics" or a tab of the R adjusted function of MARS.
}
\description{
`result_MARS()` Return plots or tab depending of the "tab" argument of the R adjusted of abundance/metric value models
at each scales for gam with the corresponding signifiance if it can be tested.
}
\details{
This function is meant to be used with the tab generated by formodel() merged with the tab of your abundance sampling.
}
\author{
Pierre-Gilles Lemasle <pg.lemasle@gmail.com>
}
