% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_lm.R
\name{result_lm}
\alias{result_lm}
\title{Computed the R adjusted of the abundance/metric value models at each scales with the lm method.}
\usage{
result_lm(temp, dist, metrics, Sp_Name, tab = T, plots = T,
  individual_plots = F)
}
\arguments{
\item{temp}{The tab generated by formodel() function merged by the id of sampling site
with the tab of your abundance sampling.}

\item{dist}{Vector of scales you choosed in Chloe when you created the ascii whith Cloe}

\item{metrics}{Vector of metrics of 1 or more you choosed in Chloe when you created the ascii whith Chloe}

\item{tab}{Argument "T" or "F" defining if you want to return plot (F) or a tab (T).}

\item{sp_name}{Optionnal argument containing the name of the species you want to see analyse, the name
must match its column name contained in the "carab" tab}
}
\value{
Return a list of plot(s) of the length of "metrics" or a tab of the R adjusted function of lm
and their signifiance.
}
\description{
`result_lm()` Return plots or tab depending of the "tab" argument of the R adjusted of abundance/metric value models
at each scales for lm, gam and MARS with the corresponding signifiance if it can be tested.
}
\details{
This function is meant to be used with the tab generated by formodel() merged with
the tab of your abundance sampling.
}
\author{
Pierre-Gilles Lemasle <pg.lemasle@gmail.com>
}
