% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_Cor.R
\name{RV_COR}
\alias{RV_COR}
\title{Computed the RV coefficient between the metrics.}
\usage{
RV_COR(Varia_paysage_multi, metrics, dist)
}
\arguments{
\item{Varia_paysage_multi}{Tab generated by formultivariate()}

\item{metrics}{Vector of  metrics  you choosed in Chloe during your analysis in Chloe}

\item{dist}{Vector of scales you choosed during you analusis in Chloe}
}
\value{
Return a correlogram of the uncorrelated metrics of the Chloe analylis.
}
\description{
`RV_COR()` Return a correlogram of the uncorrelated metrics of the Chloe analylis.
}
\details{
This function is meant to be used with the tab generated by formultivariate()
The RV coefficient is meant to computed the correlation between two tabs.
In this case metric's values at each scale is considered as tab. So a RV coefficient is computed
for each pairwise metrics.
}
